Fixes #212 refactor ArtifactPath from internal to certified
