+++ certification/internal/shell/operatorsdk.go
@@ -7,10 +7,11 @@ import (
 	"io/ioutil"
 	"os"
 	"os/exec"
+	"path/filepath"
 	"strings"
 
 	"github.com/redhat-openshift-ecosystem/openshift-preflight/certification/errors"
-	fileutils "github.com/redhat-openshift-ecosystem/openshift-preflight/certification/internal/utils/file"
+	certutils "github.com/redhat-openshift-ecosystem/openshift-preflight/certification/utils"
 	"github.com/redhat-openshift-ecosystem/openshift-preflight/cli"
 	log "github.com/sirupsen/logrus"
 )
@@ -134,7 +135,7 @@ func (o OperatorSdkCLIEngine) BundleValidate(image string, opts cli.OperatorSdkB
 }
 
 func (o OperatorSdkCLIEngine) writeScorecardFile(resultFile, stdout string) error {
-	scorecardFile := fileutils.ArtifactPath(resultFile)
+	scorecardFile := filepath.Join(certutils.ArtifactPath(), resultFile)
 
 	err := ioutil.WriteFile(scorecardFile, []byte(stdout), 0644)
 	if err != nil {
