+++ cmd/check_operator.go
@@ -12,6 +12,8 @@ import (
 	"github.com/redhat-openshift-ecosystem/openshift-preflight/certification/formatters"
 	"github.com/redhat-openshift-ecosystem/openshift-preflight/certification/runtime"
 	"github.com/spf13/cobra"
+
+	certutils "github.com/redhat-openshift-ecosystem/openshift-preflight/certification/utils"
 )
 
 var checkOperatorCmd = &cobra.Command{
@@ -52,19 +54,17 @@ var checkOperatorCmd = &cobra.Command{
 			return err
 		}
 
-		 artifactpath := ArtifactPath()
-
 		// create the results file early to catch cases where we are not
 		// able to write to the filesystem before we attempt to execute checks.
 		resultsFile, err := os.OpenFile(
-			filepath.Join(artifactpath, resultsFilenameWithExtension(formatter.FileExtension())),
+			filepath.Join(certutils.ArtifactPath(), resultsFilenameWithExtension(formatter.FileExtension())),
 			os.O_WRONLY|os.O_CREATE|os.O_TRUNC,
 			0600,
 		)
 
 		if err != nil {
 			return err
-		}		
+		}
 
 		// also write to stdout
 		resultsOutputTarget := io.MultiWriter(os.Stdout, resultsFile)
